Added check in ViewerPanel destructor for null AirfoilStructs on deletion
The airfoils are deleted in the loop just prior to this when the AirfoilListStruct is cleared, so this loop is acting more like a backup deletion for any that weren't stored correctly in the afListMembers vector. Since they are deleted beforehand, the program was segfaulting on exit when it tried to delete the airfoil structs twice.